requirements.txt
kivy
kivymd
pyrebase4
python-firebase
requests